Update the less specific route node with requested values
It has been observed that if the less specific route is deleted first
and then all the more specific routes, the less specific route doesn't
go away. For e.g.: if 1.1.1.0/24 is deleted first and later 1.1.1.1/32
in a table where there are no other routes, 1.1.1.0/24 remains in the
table.

When we delete a non leaf node, if it is a bucket we do not update the
prefix length, flags and label values. If these values do not match with
those of other nodes in the same bucket, then the bucket will remain
forever. Hence, update the node with the values in the request before
traversing the tree down.

Change-Id: I71778a83093a04c462b40e6c7e05b1d993f3373e
Closes-BUG: #1563734