Set skb network header for non-tunneled packets
When somebody does a tcpdump or an equivalent operation of listening
on an interface for all protocol packets, linux kernel prints protocol
xyz is buggy if the network header is positioned at an offset that is
not sane, for e.g.: network header offset is lesser than the data offset
(which is where the data starts).

vRouter is not setting network header offsets for packets that don't
undergo GRO or for non-tunnel packets. If such packets egress an
interface and if a tcpdump like application is running on that interface,
then we will see this issue.

Set skb network header to packet network header for such packets

Change-Id: I82f7550722798e53c94092788b5135603f6045c8
Closes-BUG: #1564242