ECMP without flow
Currently ECMP does not work if the packets do not transit the flow table
(i.e.: if there is no policy that guides the packet). The requirement for
flow lookup comes from the fact that flows need to follow the same path
(also called flow stickiness) regardless of existing destinations leaving
and new destinations entering, which affects the list of ECMP destinations,
and thus the distribution. To workaround that problem, we maintain an index
in the flow entry that points to the correct end point in the distribution
list.

For non-policy looked up packets, requirement of following the same path in
case of transitions is dropped.

Change-Id: Ib7688c9dcac522025196334f469844c79ebdf095
Partial-BUG: #1566650