Display used flow table count and bridge table count in "vrouter" utility
The uses flow table count and bridge table count is not displayed as
part of any utility. "vrouter" utility is extended to display this
count. This can be further extended to display all the Vrouter table
count like mpls, nh etc. As of now, bridge table and flow table count
along with its over flow table is displayed as part of "vrouter --info"

closes-bug: #1569791

Change-Id: I2e54fb8e4fd89a15ecafdf5551f022448728a5f3