Incasing the overflow table size of Flow and Bridge tables
Currently the overflow table sizes of flow and bridge are kept at 8k and
4k respectively as constant, irrespective of the main flow table size.
Now this is changed to 20% of the main table. If user specifies over
flow entries as module parameter, the value is untouched and left as is.
If user specifies only the main table entries, then over flow table
entries are again calculated as 20% of main table. While doing the
calculation, it is adjusted to upper bound 1k.

Change-Id: I45bb463013f850824af050ef9fb6d86118569f52
closes-bug: #1573381