Create a new flow if a packet hits a delete marked flow
When Agent deletes a flow, the flow entry is not immediately deleted in
Vrouter. Vrouter uses work queues and rcu callbacks to streamline the
deletion. After these asynchronous callbacks, flow entry is treated as
deleted. If there are any packets hitting this flow entry before marking
as "not active" those packets are dropped. On a loaded system with huge
number of flows, these call backs take a considerable time(in the worst
case as much as 30 ms). Packets hitting this flow entry will be dropped
for all this time.

To avoid this issue, when Agent deletes the flow entry, entry is "marked
deleted". A delete marked entry is no more treated as active entry,
hence flow lookup does not return this entry. As a result, a new flow
entry gets created. All the subsequent packets hit the new entry. "flow"
utility displays this particular entry as "M".

Change-Id: Ia40c5a943ea2d06fa01fa3a5a60d52cba1ecfedb
closes-bug: #1569786