Before initiating eviction, wait for the reverse flow to move from HOLD
If the reverse flow action is HOLD, it is possible that we might not
flush the packets that are in the hold queue when we do eviction.
There is also an unidentified case where such flows remain in the
table in the eviction marked state.

Also, free the vrouter operation request structure once the request
is done.

Change-Id: I75d54367c7f7a3b32c3a09332be4f01a64e06d2a
Partial-BUG: #1570203