Disabling GRO when mirroring is enabled
When the packets are received on Fabric interface, these are subjected
to GRO before transmitting them to Tap interface. If Tx port mirroring
is enabled on Tap interface, as of now, mirroring is applied after GRO
of the packets. If the mirroring server is on another compute node, these
mirrored packets have GSO enabled. Vrouter makes use of Linux GSO routines
and these GSO routines expect that skb in skb_list of head_skb does not
contain any linear data (skb_headlen should be zero). Due to GRO the
skb_headle of some skb's in skb_list contains linear data resulting in
GSO routiness hitting a BUG_ON.

To over come this, the GRO needs to be applied post mirroring. To
enabled GSO on the mirrored packets the skb's gso_len also need to be
supplied. Instead of this fix, GRO is disabled if Tx port mirroring is
enabled on Tap interface as stop gap fix till the complete fix is in
place.

Change-Id: I7dd86e1bd90fef60efa4c9dbac78e853952c4fdc
closes-bug: #1577473