Manipulating the TTL of the Packet
Right now the TTL of the packet is not overwritten by Vrouter. It is
only decremented like a hop, for the required packets. But BGP packets
in VM (due to BGP as service), can come to Vrouter to with ttl of 1. In
a multi hop environment, we require this ttl to be more than 1. For this
purpose flow entry is added with another ttl field and if Agent sets
this, vrouter unconditionally sets this ttl in the packet and calculates
the checksum again.

For the time being, this is only done for Nat'ed packets.

Change-Id: I53a192666731f4c3662e3791006dd7294bccf116
partial-bug: #1567586