Inrease the Trap size of the truncated packet while Trapping to Agent
Currently for certain Trap messages the packet is trncated to 128 bytes.
if the packet is truncated to 128 bytes, Agent does not get to see the
transport header depending on the encapsulation of the packet. The
ICMP packet being received with Vxlan encap  is one such scenario where
Vrouter creates a flow with right Transport ports but as Agent does not
see the correct ports it handles that as different flow. This leads to
Hold entries being left in Vrouter.

To avoid this, instead of using the standard 128 bytes as Truncation, it
is calculated using the headers length.

Change-Id: Ifba2e7af4f1eb2310593a14a81350446273b2875
closes-bug: #1588623