Fixes vif vRouter segment fault with bad arguments
When vif is used with some invalid combinations of arguments it will
core dump with a segmentation fault. For example if vif -list is done
a core dump will happen. This is due to calling strtoul with a NULL
string. Adding a new procedure that will check if the string is NULL.

Change-Id: I5015c07e76a4434498229862f5172c2802ae99e9
Closes-Bug: #1590964