Pull the SCTP header in the ICMP error headers to Head SKB
When the packet is received on Fabric interface right now the inner
header are pulled if they are TCP, UDP, ICMP and SCTP. In case of ICMP
error, the IP header that caulsed the ICMP error is also pulled. While
pulling this payload IP header the transport heder is pulled onl if it
is TCP or UDP or ICMP. But if the ICMP error is caused by SCTP packet,
this SCTP transport header which is part of ICMP payload is not
pulled.  This is resulting in Vrouter seeing a different SCTP ports
which are not valid and hence a flow new flow is created thopugh the
original flow exists for these SCTP flows. As Agent is seeing the
correct packet it does not update the newly created wrong flow and is
not moved out of Hold state. Agent eventually removes this from as part
of its aging cycle.

As a fix, ICMP pauyload SCTP transport header also is pulled to head
SKB.

Also as part of above fix, it is observed that transport header checksum
filed is wrongly pointing to ICMP error payload's transport header
checksum rather to ICMP checksum itself in case of ICMP errors. So this
has been modified so that th_csum always points to right transport
header checksum.

closes-bug: #1588925
o
Change-Id: I4a843224f92e61263a4598074ff9baa1a864eb78