For mirroring, ECMP index should not come from parent flow
In the case of mirror to an ECMP destination, we can not take the
ECMP index that is set in the flow of the parent (the flow/packet
that is mirrored). Hence, set the ECMP index to -1.

Also, there isn't a flow that maps to the mirror packet. So,
extract all the information that is needed from the parent flow and
set the flow index in the forrwarding metadata to -1.

Change-Id: Ib3a0f77490153fbe69c35eaa201fbf40b6897d4d
Closes-BUG: #1566650