If NH doesn't have policy, do not create flows even if the source is an ECMP
For providing stickiness in the case of ECMP, vRouter explicitly used to
create flows even if the corresponding NextHop didn't have the policy bit
set. This now is a wrong behavior when we are supporting ECMP without
flows. Hence, do not create such flows, if the nexthop doesn't have
policy enabled.

Change-Id: I48c28150b75e7c78d4d093de327caf98a19ff420
Partial-BUG: #1566650