.1p based QOS marking for L2 packets
For packets that are not IPv4 or IPv6, qos remarking will be based on
802.1p information present in the packet. For supporting this feature,
we do the following:

. save the priority information in the packet before untagging the
  packet, if it comes in the vRouter vlan interface
. use this saved information to lookup in the qos configuration table
  to determine the forwarding class, as is the case with DSCP based
  lookups
. save and restore 802.1p information in the forwarding metadata, if
  and when the packet is queued somewhere

Also, fix the incremental checksum code that comes into play when tos
bits in the IPv4 header is changed. There is a problem with the current
code where the diff is shifted by 8 bits to reflect the position of TOS
bits in the IPv4 header. We should do this shift to original values
rather than the diff.

Change-Id: Iade0adae4163a3dd5e4ad9ae41cb91c6bcbb8c0d
Closes-Bug: #1603340
Closes-Bug: #1603864