Removing ref counting for Mirror entry
Flow and VIF hold Mirroring index rather reference to mirroring entry.
Though only index is used, the ref counting is used to delete the
mirroring entry when Agent deletes it. This is unncecessary as no other
data structure is holding a reference to this. Also the current ref
counting is not atomic and is leading to wrong ref counting, leading to
mirroring entry's wrong deletion.

As a fix, the ref counting is remvoed for mirroring. When Agent deletes
the entry, it is removed from Mirroring database. While mirroring if
mirroring entry is not found the packet just does not get mirrored.

closes-bug: #1605874

Change-Id: I0381fc9e0f8106b3010ef7681aed17e04c7a90cb