Avoiding the Ecmp trap mesages to Agent
Currently Vrouter has two different Trap messages to Agent incase of
Ecmp.

1) When the packet is transmitted from VM to an Ecmp destination, and
if the Ecmp destination is not available in Ecmp Nexthop. In this case a
trap to Agent results in chosing a new Ecmp destination.

2) When the packet is received from an Ecmp source, which is a valid
source, but not the one which Vrouter is transmitting to.  In this case,
Agent manipulates the the forward flow to point to latest Ecmp source.

These Trap messages are not rate limited, and if the packet rate is
high, Vrouter ends up in sending too many Trap messages to Agent.

To avoid this, both the scenarios are handled in Vrouter itself. The
Ecmp index in Flow entry is not added by Agent any more. When the first
packet hits the Ecmp nexthop, the component nexthop is chosen by using
the packet key hash value. While chosing the nexthop it is ensured
that the nexthop is a valid one. This index is manipulated in the flow
entry by Vrouter itself. If there is a change in the Ecmp source the
forward flow's index is also manipulated by Vrouter.

Change-Id: Iece9c7467bcfb4d8d1084ed7ea1a5924d2176391
partial-bug: #1573265