Update prefix len in Mtrie bucket as well
Right now, if mtrie already contains more specific entries and if a less
specific entry is added, only non-matching leaves are updated with less
specific entry's prefix len. The matching buckets also need to be
updated with prefix len. Not updating the prefix len in buckets, results
in, not collapsing the tree, if these less specific entries are deleted.

For example, in an empty tree, 8.1.1.10/32 -> nh1 is added. This results
in, at level 1 bucket 8, level 2 bucket 1, level 3 bucket 1 to contain
prefix len as 0 and level 4 bucket 10 to contain prefix len as 32. The
other leaves in level 4 (like 8.1.1.1 till 8.1.1.255 except 10) also
contain prefix len as 0. After this 8.1.1.0/24 -> nh2 is added.
Currently the prefix len of 24 is updated only in the leaves like, 8.1.1.1
till 8.1.1.255. Ideally the prefix len 24 should be updated in level 3
bucket 1 as well. Not updating this, leads to issues while collapsing
level 3 buckets. As a fix, prefix len is updated in this as well.

Change-Id: I837a32ecc8d6c15fa8644028d63ef38f2259e3ab
closes-bug: #1605748