Send packets to a specific hardware queue
With the existing support in vRouter for QOS, one can specify
the queue to which a packet should land up in addition to QOS
remarking based on QOS parameters of the inner packet (packet
from the VM). To make sure that this packet really lands up
in the intended queue, the platform specific portions need to
make use of the queue information and direct the packet to the
specific queue. In the case of Linux kernel, this plumbing is
basically done through the queue_mapping field of the skbuff.

Specific makes of Intel and other NICs support configuration
of hardware queue parameters through the DataCenterBridging
(DCB) feature. These parameters include configuration of
priority groups, configuring bandwidth limitation for these
priority groups and specifying a scheduling algorithm for
handling these queues. While these features are activated if
one uses lldpad daemon, it mandates the use of DCB on both
sides of the wire, which for a person just wanting to use the
bandwidth control feature might be an overkill. Hence, add
support to the 'qosmap' utility to make use of these features,
in case on wants to use these features in a stand-alone mode.

Change-Id: I8f04721cf860cdb0850f128c87282a116625e719
Partial-Bug: #1617938