Respond with stitched MAC for unicast ARP request
When an unicast ARP request is received on fabric interface of a compute
node, from BMS behind QFX, if MX is in ecmp, the source IP lookup might
point to subnet route pointing to Ecmp nexthop. This is because there
would not be any host route for BMS in inet table. This results in Vrouter
responding with Vhost mac address though the destination IP address is
stitched. As such this behaviour is to ensure that Routing is forced if
the packet is from Ecmp source, though destination is in same subnet.
But this behaviour creates issues to BMS behind QFX, if BMS refreshes
ARP with unicast ARP request.

As a fix, the multicast ARP requests from Ecmp source on fabric
interface are dropped. If unicast ARP requests and if the destination
mac address of the ethernet packet is stitched mac, the ARP reply is
sent with stitched mac. If destination mac address does not match with
stitched mac, that ARP request is not processed by Vrouter.

Change-Id: I58b91e410ecee809264d8914f75816cfb584ce17
closes-bug: #1594165