When in a softirq context, memory allocation should be atomic
Eviction of flows based on TCP state machine happens in the softirq
context. The process of evicting a flow depends on the kernel work
queues, for which allocation of memory needs to happen. Under such
cases, the allocation should not wait. Hence, pass GFP_ATOMIC as the
allocation flag when in softirq context to indicate that the path is
not willing to block.

Change-Id: I2c56203c2666a981ff3218268611d95e18281151
Closes-Bug: #1618375