Distribute mirroring packets to Ecmp destination
Incase of dynamic mirroring, udp header is added on top of mirrored
packet. If the mirroring destination is an Ecmp destination, ecmp
distribution is happening using the newly added udp heder. This results
in no distribution with in Ecmp, as all mirrored packets are going to
contain the same IP/UDP header resulting in same hash calculation.

As a fix, for a better distribution, we need to hash on the VM's packets. For this
reason, hash  is calculated before encapsulating the
packet with UDP header itself, rather calculating while processing Ecmp
nexrhop. The hash is calculated by identifying whether the mirroring
destination is pointing to Ecmp nexthop. This nexthop is retrieved by
doing an explicit route lookup.

Change-Id: I06f2fd0dee79ade11696bc2b2d9c3c34c374ff65
closes-bug: #1616822