Do not add 1 to the queue in the common code
Linux treats queue id 0 specially and hence starts its queue numbering
from 1 (for modules external to the queueing subsystem) and hence the
queueing subsystem subtracts 1 from the queue id stored in the socket
buffer. To take care of this behavior, we had added 1 to the queue id
in the common part (os independent part) of vRouter code. This is not
a desirable behavior since other platforms might need the original
value. Hence, add 1 to the queue id in the linux specific part of
vRouter code.

Change-Id: Iad52518bbcc03f3a4b8d4292c7fc6faadb4da057
Partial-Bug: #1617938