In HashTables, handle the error case if deferred work entry creation fails
Right now, if the deferred work entry creation fails, we do not take any
ation. This can result in, Hash bucket never being cleaned up, as there
is no work entry scheduled at all.

As a fix, reset the work entry 'scheduled' variable, so that, it gets
scheduled when the next hash entry is attemped for deletion in that
bucket

Change-Id: Ia66bd340177253ac1267aa667521c26838487dbe
partial-bug: #1618375