Support for Ecmp configured hash in Vrouter
The hash parameters that are required to calculate hash value to chose
the Ecmp candidate NH are placed in Ecmp composite nexthop. After the
flow key is formed, using the configured bit fields, flow key values are
reset to zero if they are not required. Hash is calculated after
resetting the required fields.

partial-bug: #1616814

Change-Id: If674a35c07782ca227394a901760840dde5b9359