Validate MAC size before accessing MAC in bridge table configs
Currently when Agent sends a bridge table add/delete/dump request
it sends MAC address in "route" sandesh structure. Vrouter is accessing
the rtr_mac field from sandesh decoded structure without validating if
rtr_mac is passed by Agent or not. If Agent does not send this field,
Vrouter crashes as we are accessing NULL pointer.

As a fix, the mac size is validated before accessing the mac.

Change-Id: I89f03e0f5a95b051361f3b242bbeef891fe93144
closes-bug: #1623896