Dont display last displayed route in rt dump
Mtrie dump, dumps 32 routes per one iteration. "rt" utility sends last displayed
route in rt_marker in the next iteration. Vrouter searchs till rt_marker and
starts displaying the routes again from rt_marker. This results in rt_marker
being displayed in previous and as well as current iteration.

As such rt_marker_plen is not being used while searching the rt_marker. So the
usage of this is removed in Vrouter. Sandesh file is not changed so that
Agent does not have to undergo any further changes.

Change-Id: Ie6eaf39f6fe40faf88cf4175ad594fa62fb8b582
closes-bug: #1614847