Identify IPv6 diagnostic packets
Diagnostic packets are identified by two characteristics of the
packet

. wrong layer 4 checksum
. layer 4 checksum of 0xFFFF

When vRouter receives a packet and if it identifies that the packet
has a checksum of 0xFFFF, it tries to verify the checksum. If the
checksum turns out to be wrong, it marks the packet as a diagnostic
packet.

Before this change, we only used to verify checksum for IPv4 packets.
Extend that logic to IPv6 packets too.

Change-Id: Ida650cf02ad2bcee1fc1716b6903544068947a37
Partial-Bug: #1613999