Correct wrong pull of ICMP error packet
While pulling the packet to Head SKB in MplsOGre encapsulation
for ICMP error packet, the length to pull is calculated wrong.
An unnecessary subtraction of length of 8 bytes, has caused not
to pull the required transport header. This is resulting in
wrong processing of inner transport header of ICMP error packet, leading
to wrong flow being created in Vrouter, with spurious UDP ports.
Agent as such creates right flow and fails to write the flow to Vrouter
as the Vrouter returns flow key mismatch error. This results in too
many flows being left in Hold state as Agent fails to delete the
flow.

As a fix, the correct transport header pull length is maitained.

Change-Id: I1b935edc2ffd353f4b418c01718a0945971615cc
closes-bug: #1627033