If a pclone fails, trap the original packet
In the process of indicating a flow miss to the agent, if a pclone
fails, vRouter returns without unsetting the NEW flag, a flag that
indicates that the datapath is in the process of creating a new flow.
The side effect of not unsetting this flag is that agent can not
modify the entry and this leads to perennial HOLD flows in the
system.

To fix this problem, if the pclone fails, trap the original packet
to the agent without holding it in the HOLD queue. While this will
mean that the first packet does not reach the destination, it is
a better alternative to not trapping the packet and wait for agent
to do the audit after a minute or so. Such trapped packets are
captured under a new counter "Original Packet Trapped" in the
dropstats output.

Change-Id: I34c19abf935d9b06f55e875b76a4859350743c2b
Partial-Bug: #1628175