Maintain Packet Drop stats per interface
The dropstats need to be available per interface. Incrementing
atomically the single counter across many CPUs might attract
significatnt delay. Alternative approach of allocating memory for
every dropstat for every CPU will be large chunk of memory
considering large number of interfaces, number of dropstats,
number of CPUs and size of every counter.  To normalise both the
requirements, one set of dropstats of 8 byte size and another set
of dropstats of 1 byte per every cpu is allocated. The later is
incremented without any contention as it is per cpu. When the
one byte counter reaches its max value, it is added to the
64 bit counter atomically. This results in decreasing the delay
as well decresing the memory requirement, which will not be
contiguous as well. If for some reasong, the per cpu memory
allocations fails, we continue to work on the 8 byte counters
by incrementing atomically.

The regular "dropstats" command takes "core" as optional argument
as all the dropstats are currently maintained per core. With the
above changes, we will not be maintaing the stats per core, so the
option "core" is dropped from "dropstats" command. A new option
"--intf" is added to fetch the statistics per interface. Not providing
the interface fetches the statistics per vrouter as before.

Also, there are good number of statistics which are not in use. These have
been removed from counting and display.

Change-Id: Ie77d2385bcb1be77b3fb4f43341996defb213e92
partial-bug: #1515663