Enable Vrouter Vtests
Review request https://review.opencontrail.org/12698 has broken the
vrouter tests and the tests are disabled subsequently because of this.
Now the vr_flow_response sandesh message is added and the flow source ip
and destinatio ip is correctly added to xml file. This passes the tests.
The tests are reenabled too.

Change-Id: I99be64082eb3e85ea13529dd35cac14802cb3487
partial-bug: #1479295