Change in port mirroring tap locations
For all types of virtual interfaces, the location from where we tap
packets for RX port mirroring is in a common function where all
the common virtual processing happens. For a vlan based sub-interface,
this means that the port mirroring happens post untagging the packet
and hence this is not a correct behavior.

Hence, move the tap locations to the lowest layers of vif receive.

Also, during transmission, since the tap is in the lowest common
layer, sub interface mirroring does not work. Hence, move the taps
to the highest layer of vif transmit.

Also, add taps for vhost and physical interfaces.

Change-Id: Ie42bfcedde2eba1167fa49e7ecd627b7762363e9
Closes-Bug: #1630889